Ticket #396: openib BTL subnet fix Merged the attached diff file, which closes trac:396
Submitted by: gleb
Reviewed by: jsquyres
Merged by: Tim Mattox

This commit was SVN r11755.

The following Trac tickets were found above:
  Ticket 396 --> https://svn.open-mpi.org/trac/ompi/ticket/396

+++ ompi/mca/btl/openib/btl_openib_endpoint.h
@@ -41,7 +41,7 @@ struct mca_btl_openib_frag_t;
 
 struct mca_btl_openib_port_info_t {
     uint32_t mtu;
-    uint16_t subnet; 
+    uint64_t subnet; 
 };
 typedef struct mca_btl_openib_port_info_t mca_btl_openib_port_info_t;
 
@@ -88,7 +88,7 @@ struct mca_btl_openib_rem_info_t {
     uint32_t                    rem_psn_lp; 
     /* Remote processes port sequence number (Low and High) */ 
    
-    uint16_t                    rem_subnet; 
+    uint64_t                    rem_subnet; 
     /* subnet of remote process */     
 
     /* MTU of remote process */
